Contests/PlaidCTF/2025/TaLeS from the crypt/requirements.txt
pyOpenSSL==24.1.0
pycryptodome==3.20.0
